requirements.txt
# --- requirements.txt (Dependencies) ---
fastapi
uvicorn[standard]
sqlalchemy
pymysql
celery
redis
pillow
requests
python-dotenv
aiofiles